rename proposal module
