+++ apps/web/src/modules/proposal/components/ProposalActions/VoteStatus/Pending.tsx
