+++ apps/web/src/modules/proposal/components/ProposalActions/VoteStatus/Vote.tsx
