+++ apps/web/src/modules/proposal/components/ProposalActions/VoteStatus/VoteModal.tsx
