+++ apps/web/src/modules/proposal/components/ProposalCard.css.ts
