+++ apps/web/src/modules/proposal/components/ProposalCard.tsx
