+++ apps/web/src/modules/proposal/components/ProposalDescription/ProposalDescription.tsx
