+++ apps/web/src/modules/proposal/components/ProposalDescription/index.ts
