+++ apps/web/src/modules/proposal/components/ProposalDetailsGrid/ProposalDetailsGrid.tsx
