+++ apps/web/src/modules/proposal/components/ProposalDetailsGrid/Tile.css.ts
