+++ apps/web/src/modules/proposal/components/ProposalDetailsGrid/Tile.tsx
