+++ apps/web/src/modules/proposal/components/ProposalDetailsGrid/index.ts
