+++ apps/web/src/modules/proposal/utils/isProposalOpen.test.ts
