+++ apps/web/src/modules/proposal/utils/isProposalOpen.ts
