+++ apps/web/src/modules/proposal/utils/isProposalSuccessful.test.ts
