+++ apps/web/src/modules/proposal/utils/isProposalSuccessful.ts
